Add the initial development Talos configuration
Add the initial patch files for the various Talos machines, with
overrides for the type of node as well as the individual nodes as well.
Also start the supporting of individual patches which can be deployed
out to all nodes on a cluster individually, or incorporated in to
machine configurations for redeployment of a node.